Reorganization of views
